project/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
idna==2.10
psycopg2==2.8.6
pycodestyle==2.7.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
